renaming components/dataset_filter.py to components/filter.py
