+++ src/datumaro/components/hl_ops.py
@@ -11,7 +11,7 @@ from typing import Dict, Optional, Type, Union
 from datumaro.components.dataset import Dataset, IDataset
 from datumaro.components.dataset_item_storage import DatasetItemStorageDatasetView
 from datumaro.components.dataset_base import Transform
-from datumaro.components.dataset_filter import XPathAnnotationsFilter, XPathDatasetFilter
+from datumaro.components.filter import XPathAnnotationsFilter, XPathDatasetFilter
 from datumaro.components.environment import Environment
 from datumaro.components.exporter import Exporter
 from datumaro.components.launcher import Launcher, ModelTransform
